% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phr_translations.R
\name{get_phr_translations}
\alias{get_phr_translations}
\title{Get current translation json loaded in the phr_environment}
\usage{
get_phr_translations()
}
\description{
Get current translation json loaded in the phr_environment
}
