% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_greater_than_one_selection.R
\name{is_greater_than_one_selection}
\alias{is_greater_than_one_selection}
\title{Check if a Select Multiple Input has More Than One Selection}
\usage{
is_greater_than_one_selection(
  select_multiple_input,
  select_multiple_col = NULL
)
}
\arguments{
\item{select_multiple_input}{A data frame, tibble, or a character vector containing the relevant column or values.}

\item{select_multiple_col}{A character string specifying the name of the `select_multiple` column to be checked (if a data frame is provided).}
}
\value{
A logical vector where `TRUE` indicates that a row or vector element has more than one selection and `FALSE` otherwise.
}
\description{
This helper function checks if rows in a `select_multiple` column or a vector have more than one selection (response) by splitting responses based on whitespace delimiters.
}
\examples{
# Example dataset
df <- data.frame(
  select_multiple_column = c(
    "option_a option_b",
    "option_c",
    "option_d option_e option_f option_g",
    ""
  )
)

# Check with a data frame
is_greater_than_one_selection(df, "select_multiple_column")
# Output: TRUE, FALSE, TRUE, FALSE

# Check with a vector
vec <- c("option_a option_b", "option_c", "option_d option_e option_f option_g", "")
is_greater_than_one_selection(vec)
# Output: TRUE, FALSE, TRUE, FALSE

}
