% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_init_translations.R
\name{phr_init_translations}
\alias{phr_init_translations}
\title{Initialize saved json translation files}
\usage{
phr_init_translations()
}
\description{
Useful for development when translation files are updated.
}
