% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_load_translations.R
\name{phr_load_translations}
\alias{phr_load_translations}
\title{Load translation dictionaries from JSON files in inst/app/www/i18n/}
\usage{
phr_load_translations()
}
\description{
Load translation dictionaries from JSON files in inst/app/www/i18n/
}
