% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_reload_translations.R
\name{phr_reload_translations}
\alias{phr_reload_translations}
\title{Reload translations from JSON files}
\usage{
phr_reload_translations()
}
\description{
Useful for development when translation files are updated.
}
