% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_text_to_key.R
\name{phr_text_to_key}
\alias{phr_text_to_key}
\title{Convert a text string to a translation key format}
\usage{
phr_text_to_key(text)
}
\arguments{
\item{text}{The text string to convert}
}
\value{
A lowercase key with underscores instead of spaces
}
\description{
Convert a text string to a translation key format
}
