% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_validate_all_character.R
\name{phr_validate_all_character}
\alias{phr_validate_all_character}
\title{Validate Entire Vector is Character or Coercible to Character}
\usage{
phr_validate_all_character(
  x,
  allowed_values = NULL,
  origin = NULL,
  hint = NULL,
  soft
)
}
\arguments{
\item{x}{Vector or data frame column to test.}

\item{allowed_values}{Optional character vector of allowed values. If provided,
validation will fail if any element of `x` is not found within this set.}

\item{origin}{Optional name of the originating function (for tracing errors).}

\item{hint}{Optional corrective hint.}

\item{soft}{Logical; if TRUE, issues a warning on failure; if FALSE, throws an error.}
}
\value{
Invisibly returns TRUE if valid; otherwise triggers `phr_error()` or `phr_warning()`.
}
\description{
Checks that all values in a vector (or column) are character strings or can be
represented as character safely. Optionally enforces that all values belong to a
specified set of allowable options.
}
