% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_validate_all_date.R
\name{phr_validate_all_date}
\alias{phr_validate_all_date}
\title{Validate Entire Vector is Date or Coercible to Date}
\usage{
phr_validate_all_date(x, origin = NULL, hint = NULL, soft)
}
\arguments{
\item{x}{Vector or data frame column to test.}

\item{origin}{Optional name of the originating function.}

\item{hint}{Optional corrective hint.}

\item{soft}{Logical; if TRUE, issues a warning on failure; if FALSE, throws an error.}
}
\value{
Invisibly returns TRUE if valid; otherwise triggers `phr_error()` or `phr_warning()`.
}
\description{
Checks that all values in a vector (or column) are Date, POSIX, or can be safely
converted to a Date representation using common formats.
}
