% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_validate_all_numeric.R
\name{phr_validate_all_numeric}
\alias{phr_validate_all_numeric}
\title{Validate Entire Vector is Numeric or Coercible to Numeric}
\usage{
phr_validate_all_numeric(x, origin = NULL, hint = NULL, soft)
}
\arguments{
\item{x}{Vector or data frame column to test.}

\item{origin}{Optional name of the originating function.}

\item{hint}{Optional corrective hint.}

\item{soft}{Logical; if TRUE, issues a warning on failure; if FALSE, throws an error.}
}
\value{
Invisibly returns TRUE if valid; otherwise triggers `phr_error()`.
}
\description{
Checks that all values in a vector (or column) are numeric or can be safely
converted to numeric without introducing NAs.
}
