% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_validate_character.R
\name{phr_validate_character}
\alias{phr_validate_character}
\title{Validate Character Input (Strict)}
\usage{
phr_validate_character(x, origin = NULL, hint = NULL, soft)
}
\arguments{
\item{x}{Object to test.}

\item{origin}{Optional name of originating function.}

\item{hint}{Optional corrective hint.}

\item{soft}{Logical; if TRUE, issues a warning on failure; if FALSE, throws an error.}
}
\value{
Invisibly returns TRUE if valid; otherwise triggers `phr_error()` or `phr_warning()`.
}
\description{
Ensures that `x` is a character vector (not factor, numeric, or logical).
}
