% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_validate_choice.R
\name{phr_validate_choice}
\alias{phr_validate_choice}
\title{Validate Choice Membership}
\usage{
phr_validate_choice(x, choices, origin = NULL, soft)
}
\arguments{
\item{x}{Value or vector to test.}

\item{choices}{Character vector of allowed values.}

\item{origin}{Optional name of the originating function.}

\item{soft}{Logical; if TRUE, issues a warning on failure; if FALSE, throws an error.}
}
\value{
Invisibly returns TRUE if valid; otherwise triggers `phr_error()` or `phr_warning()`.
}
\description{
Ensures that the input value(s) belong to a defined set of allowed options.
}
