% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phr_validate_factor.R
\name{phr_validate_factor}
\alias{phr_validate_factor}
\title{Validate Factor Input}
\usage{
phr_validate_factor(x, origin = NULL, hint = NULL, soft)
}
\arguments{
\item{x}{Object to test.}

\item{origin}{Optional name of the originating function.}

\item{hint}{Optional corrective hint.}

\item{soft}{Logical; if TRUE, issues a warning on failure; if FALSE, throws an error.}
}
\value{
Invisibly returns TRUE if valid; otherwise triggers `phr_error()` or `phr_warning()`.
}
\description{
Ensures that `x` is a factor variable.
}
